+++ src/integrationTest/java/com/mongodb/hibernate/embeddable/StructAggregateEmbeddableIntegrationTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.hibernate;
+package com.mongodb.hibernate.embeddable;
 
 import static com.mongodb.hibernate.MongoTestAssertions.assertEquals;
 import static com.mongodb.hibernate.MongoTestAssertions.assertUsingRecursiveComparison;
@@ -26,7 +26,6 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.hibernate.junit.InjectMongoCollection;
 import com.mongodb.hibernate.junit.MongoExtension;
 import jakarta.persistence.AccessType;
-import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.Entity;
@@ -48,90 +47,25 @@ import org.junit.jupiter.api.extension.ExtendWith;
 @SessionFactory(exportSchema = false)
 @DomainModel(
         annotatedClasses = {
-            EmbeddableIntegrationTests.ItemWithFlattenedValues.class,
-            EmbeddableIntegrationTests.ItemWithFattenedEmptyValue.class,
-            EmbeddableIntegrationTests.ItemWithNestedValues.class,
-            EmbeddableIntegrationTests.ItemWithNestedEmptyValue.class,
-            EmbeddableIntegrationTests.Unsupported.ItemWithNestedValueHavingNonInsertable.class,
-            EmbeddableIntegrationTests.Unsupported.ItemWithNestedValueHavingAllNonInsertable.class,
-            EmbeddableIntegrationTests.Unsupported.ItemWithNestedValueHavingNonUpdatable.class
+            StructAggregateEmbeddableIntegrationTests.ItemWithNestedValues.class,
+            StructAggregateEmbeddableIntegrationTests.ItemWithOmittedEmptyValue.class,
+            StructAggregateEmbeddableIntegrationTests.Unsupported.ItemWithNestedValueHavingNonInsertable.class,
+            StructAggregateEmbeddableIntegrationTests.Unsupported.ItemWithNestedValueHavingAllNonInsertable.class,
+            StructAggregateEmbeddableIntegrationTests.Unsupported.ItemWithNestedValueHavingNonUpdatable.class
         })
 @ExtendWith(MongoExtension.class)
-class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
+class StructAggregateEmbeddableIntegrationTests implements SessionFactoryScopeAware {
     @InjectMongoCollection("items")
     private static MongoCollection<BsonDocument> mongoCollection;
 
     private SessionFactoryScope sessionFactoryScope;
 
     @Test
-    void testFlattenedValues() {
-        var item = new ItemWithFlattenedValues();
-        {
-            item.id = new EmbeddableValue();
-            item.id.a = 1;
-            item.flattened1 = new EmbeddableValue();
-            item.flattened1.a = 2;
-            item.flattened2 = new EmbeddablePairValue1();
-            item.flattened2.a = 3;
-            item.flattened2.flattened = new EmbeddablePairValue2(4, 5);
-            item.flattened2.parent = item;
-        }
-        sessionFactoryScope.inTransaction(session -> session.persist(item));
-        assertThat(mongoCollection.find())
-                .containsExactly(new BsonDocument()
-                        .append(ID_FIELD_NAME, new BsonInt32(item.id.a))
-                        .append("flattened1_a", new BsonInt32(2))
-                        .append("flattened2_a", new BsonInt32(3))
-                        .append("flattened2_flattened_a", new BsonInt32(4))
-                        .append("flattened2_flattened_b", new BsonInt32(5)));
-        var loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithFlattenedValues.class, item.id));
-        assertEquals(item, loadedItem);
-        var updatedItem = sessionFactoryScope.fromTransaction(session -> {
-            var result = session.find(ItemWithFlattenedValues.class, item.id);
-            result.flattened1.a = -result.flattened1.a;
-            return result;
-        });
-        loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithFlattenedValues.class, item.id));
-        assertEquals(updatedItem, loadedItem);
-    }
-
-    @Test
-    void testFlattenedEmptyValue() {
-        var item = new ItemWithFattenedEmptyValue();
-        {
-            item.id = 1;
-            item.omitted = new EmbeddableEmptyValue();
-        }
-        sessionFactoryScope.inTransaction(session -> session.persist(item));
-        assertThat(mongoCollection.find())
-                .containsExactly(
-                        // Hibernate ORM does not store/read the empty `item.omitted` value.
-                        // See https://hibernate.atlassian.net/browse/HHH-11936 for more details.
-                        new BsonDocument(ID_FIELD_NAME, new BsonInt32(item.id)));
-        var loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithFattenedEmptyValue.class, item.id));
-        assertUsingRecursiveComparison(item, loadedItem, (assertion, actual) -> assertion
-                .ignoringFields("omitted")
-                .isEqualTo(actual));
-        var updatedItem = sessionFactoryScope.fromTransaction(session -> {
-            var result = session.find(ItemWithFattenedEmptyValue.class, item.id);
-            result.omitted = null;
-            return result;
-        });
-        assertThat(mongoCollection.find()).containsExactly(new BsonDocument(ID_FIELD_NAME, new BsonInt32(item.id)));
-        loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithFattenedEmptyValue.class, item.id));
-        assertEquals(updatedItem, loadedItem);
-    }
-
-    @Test
     void testNestedValues() {
         var item = new ItemWithNestedValues();
         {
-            item.id = new StructAggregateEmbeddableValue();
-            item.id.a = 1;
+            item.nestedId = new StructAggregateEmbeddableValue();
+            item.nestedId.a = 1;
             item.nested1 = new StructAggregateEmbeddableValue();
             item.nested1.a = 2;
             item.nested2 = new StructAggregateEmbeddablePairValue1();
@@ -143,7 +77,7 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
         assertThat(mongoCollection.find())
                 .containsExactly(new BsonDocument()
                         // Hibernate ORM flattens `item.id` despite it being of an aggregate type
-                        .append(ID_FIELD_NAME, new BsonInt32(item.id.a))
+                        .append(ID_FIELD_NAME, new BsonInt32(item.nestedId.a))
                         .append("nested1", new BsonDocument("a", new BsonInt32(2)))
                         .append(
                                 "nested2",
@@ -155,20 +89,21 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
                                                         .append("a", new BsonInt32(4))
                                                         .append("b", new BsonInt32(5)))));
         var loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithNestedValues.class, item.id));
+                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithNestedValues.class, item.nestedId));
         assertEquals(item, loadedItem);
         var updatedItem = sessionFactoryScope.fromTransaction(session -> {
-            var result = session.find(ItemWithNestedValues.class, item.id);
+            var result = session.find(ItemWithNestedValues.class, item.nestedId);
             result.nested1.a = -result.nested1.a;
             return result;
         });
-        loadedItem = sessionFactoryScope.fromTransaction(session -> session.find(ItemWithNestedValues.class, item.id));
+        loadedItem =
+                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithNestedValues.class, item.nestedId));
         assertEquals(updatedItem, loadedItem);
     }
 
     @Test
     void testNestedEmptyValue() {
-        var item = new ItemWithNestedEmptyValue();
+        var item = new ItemWithOmittedEmptyValue();
         {
             item.id = 1;
             item.omitted = new StructAggregateEmbeddableEmptyValue();
@@ -180,18 +115,18 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
                         // See https://hibernate.atlassian.net/browse/HHH-11936 for more details.
                         new BsonDocument(ID_FIELD_NAME, new BsonInt32(item.id)));
         var loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithNestedEmptyValue.class, item.id));
+                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithOmittedEmptyValue.class, item.id));
         assertUsingRecursiveComparison(item, loadedItem, (assertion, actual) -> assertion
                 .ignoringFields("omitted")
                 .isEqualTo(actual));
         var updatedItem = sessionFactoryScope.fromTransaction(session -> {
-            var result = session.find(ItemWithNestedEmptyValue.class, item.id);
+            var result = session.find(ItemWithOmittedEmptyValue.class, item.id);
             result.omitted = null;
             return result;
         });
         assertThat(mongoCollection.find()).containsExactly(new BsonDocument(ID_FIELD_NAME, new BsonInt32(item.id)));
         loadedItem =
-                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithNestedEmptyValue.class, item.id));
+                sessionFactoryScope.fromTransaction(session -> session.find(ItemWithOmittedEmptyValue.class, item.id));
         assertEquals(updatedItem, loadedItem);
     }
 
@@ -202,68 +137,9 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
 
     @Entity
     @Table(name = "items")
-    static class ItemWithFlattenedValues {
-        @Id
-        EmbeddableValue id;
-
-        @AttributeOverride(name = "a", column = @Column(name = "flattened1_a"))
-        EmbeddableValue flattened1;
-
-        @AttributeOverride(name = "a", column = @Column(name = "flattened2_a"))
-        @AttributeOverride(name = "flattened.a", column = @Column(name = "flattened2_flattened_a"))
-        @AttributeOverride(name = "flattened.b", column = @Column(name = "flattened2_flattened_b"))
-        EmbeddablePairValue1 flattened2;
-    }
-
-    @Embeddable
-    static class EmbeddableValue {
-        int a;
-    }
-
-    @Embeddable
-    static class EmbeddablePairValue1 {
-        int a;
-        EmbeddablePairValue2 flattened;
-
-        @Parent ItemWithFlattenedValues parent;
-
-        /**
-         * Hibernate ORM requires a getter for a {@link Parent} field, despite us using {@linkplain AccessType#FIELD
-         * field-based access}.
-         */
-        void setParent(ItemWithFlattenedValues parent) {
-            this.parent = parent;
-        }
-
-        /**
-         * Hibernate ORM requires a getter for a {@link Parent} field, despite us using {@linkplain AccessType#FIELD
-         * field-based access}.
-         */
-        ItemWithFlattenedValues getParent() {
-            return parent;
-        }
-    }
-
-    @Embeddable
-    record EmbeddablePairValue2(int a, int b) {}
-
-    @Entity
-    @Table(name = "items")
-    static class ItemWithFattenedEmptyValue {
-        @Id
-        int id;
-
-        EmbeddableEmptyValue omitted;
-    }
-
-    @Embeddable
-    static class EmbeddableEmptyValue {}
-
-    @Entity
-    @Table(name = "items")
     static class ItemWithNestedValues {
         @Id
-        StructAggregateEmbeddableValue id;
+        StructAggregateEmbeddableValue nestedId;
 
         StructAggregateEmbeddableValue nested1;
 
@@ -307,7 +183,7 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
 
     @Entity
     @Table(name = "items")
-    static class ItemWithNestedEmptyValue {
+    static class ItemWithOmittedEmptyValue {
         @Id
         int id;
 
@@ -323,7 +199,8 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
         @Test
         void testPrimaryKeySpanningMultipleFields() {
             assertThatThrownBy(() -> new MetadataSources()
-                            .addAnnotatedClass(ItemWithFlattenedPairValueAsId.class)
+                            .addAnnotatedClass(
+                                    StructAggregateEmbeddableIntegrationTests.Unsupported.ItemWithPairValueAsId.class)
                             .buildMetadata())
                     .hasMessageContaining("does not support primary key spanning multiple columns");
         }
@@ -370,9 +247,9 @@ class EmbeddableIntegrationTests implements SessionFactoryScopeAware {
 
         @Entity
         @Table(name = "items")
-        static class ItemWithFlattenedPairValueAsId {
+        static class ItemWithPairValueAsId {
             @Id
-            EmbeddablePairValue2 id;
+            StructAggregateEmbeddableIntegrationTests.StructAggregateEmbeddablePairValue2 id;
         }
 
         @Entity
